flui-engine: checked offsets in nine-slice sub-image extraction
The nine-slice region copy computed row offsets as u32 (`row * stride +
sx * 4`, `sy + sh`, `sw * sh * 4`): a large region wrapped in release
into an in-bounds read of the wrong pixels, or panicked in debug. Move
it to extract_rgba8_region, which does the arithmetic in checked usize,
reads rows through slice::get, and builds the result with
Image::try_from_rgba8.

The flui-widgets test helpers that sized RGBA8 buffers as
(width * height * 4) as usize now multiply in usize too.

+++ crates/flui-widgets/src/image/resolve.rs
@@ -275,7 +275,11 @@ mod tests {
     /// its source; two separately built images are not), so a test that wants
     /// to name a specific frame keeps the handle it published.
     fn solid(width: u32, height: u32) -> PixelImage {
-        PixelImage::from_rgba8(width, height, vec![0u8; (width * height * 4) as usize])
+        PixelImage::from_rgba8(
+            width,
+            height,
+            vec![0u8; width as usize * height as usize * 4],
+        )
     }
 
     #[test]
